stuff added

+++ src/studyplanner/View/GanttChartView.java
@@ -24,16 +24,16 @@ public class GanttChartView<X,Y> extends XYChart<X,Y> {
 
     public static class ExtraData {
 
-        public long length;
+        public float length;
         public String styleClass;
 
 
-        public ExtraData(long lengthMs, String styleClass) {
+        public ExtraData(float lengthMs, String styleClass) {
             super();
             this.length = lengthMs;
             this.styleClass = styleClass;
         }
-        public long getLength() {
+        public float getLength() {
             return length;
         }
         public void setLength(long length) {
